[route53 records] Default weight should be "-1"
When exporting r53r tfstate the weight is not set in the generated tfstate
file. Although it should be "-1" in the tfstate. That's the standard default
value.

If it's not "-1" `terraform plan/apply` wants to change all weight "" => "-1"
which is only a noop because it only changes the tfstate contents, does not
really change anything in AWS Route53.

The official documentation says:

https://www.terraform.io/docs/providers/aws/r/route53_record.html

```
Note: The weight attribute uses a special sentinel value of -1 for a default in
Terraform. This allows Terraform to distinquish between a 0 value and an empty
value in the configuration (none specified). As a result, a weight of -1 will
be present in the statefile if weight is omitted in the configuration.
```