Fix tf/tfstate generation for security groups
Adjusted the tests and both the tf/tfstate exports according to the statements
below. The tests should be improved too, for now they're ok.

Note that if it's a VPC security group, the group name is nil in the permission
set.

 EC2-Classic security groups can be:

  - security groups from the same account
  - security groups from another account
  - CIDR

If it is an EC2-Classic security group from the same account, the rule should
contain only the security group name not the group id.

Example tf export snippet: `security_groups = ["sg-1234efgh"]`

If it's an EC2-Classic security group from another account, the role should
contain the account id and the security group name, not the id.

Example tf export snippet: `security_groups = ["098765432109/sg-1234efgh"]`

VPC security groups can be:
  - security groups from the same account + same VPC
  - CIDR

A rule inside of a VPC security group can't be SGs from a different VPC or from
a differert account. Only CIDRs and SGs from the same VPC.

A rule should contain only the security group ID or a CIDR.

Example tf export snippet: `security_groups = ["sg-1234efgh"]`

The
[documentation](https://www.terraform.io/docs/providers/aws/r/security_group.html) says:

```
security_groups - (Optional)
List of security group Group Names if using EC2-Classic or the default VPC, or
Group IDs if using a non-default VPC.
```